test/data/sql -> mssql. Use default config function.
