+++ test/data/mssql/integration.query.tests.js
@@ -3,7 +3,7 @@
 // ----------------------------------------------------------------------------
 var index = require('../../../src/data/mssql'),
     queries = require('../../../src/query'),
-    config = require('./infrastructure/config'),
+    config = require('../../infrastructure/config')().data,
     expect = require('chai')
         .use(require('chai-subset'))
         .expect,
