+++ test/data/mssql/statements/createIndex.tests.js
