+++ test/data/mssql/statements/createTable.tests.js
