+++ test/data/mssql/statements/delete.tests.js
