+++ test/data/mssql/statements/insert.tests.js
