+++ test/data/mssql/statements/updateSchema.tests.js
